requirements.txt
asgiref==3.7.2
Django==4.2.6
django-ckeditor==6.7.0
django-js-asset==2.1.0
django-tinymce==3.6.1
Pillow==10.0.1
python-dotenv==1.0.0
sqlparse==0.4.4
typing-extensions==4.8.0
